move envelop playground to dedicated page
