+++ website/docs/envelop-playground.mdx
@@ -1,0 +1,5 @@
+---
+title: 'Envelop Playground'
+---
+
+<EnvelopPlayground />
