+++ website/routes.ts
@@ -11,6 +11,7 @@ export function getRoutes(): IRoutes {
           ['integrations', 'Integrations and Examples'],
           ['core', '@envelop/core'],
           ['composing-envelop', 'Sharing Envelops'],
+          ['envelop-playground', 'Envelop Playground'],
         ],
       },
       'docs/guides': {
