+++ website/src/pages/_app.tsx
@@ -1,81 +1,14 @@
 import 'remark-admonitions/styles/infima.css';
 import '../../public/style.css';
 import '../../public/admonitions.css';
-
+import * as React from 'react';
 import { appWithTranslation } from 'next-i18next';
-import {
-  chakra,
-  Code,
-  extendTheme,
-  Text,
-  theme as chakraTheme,
-  UnorderedList,
-  useColorModeValue,
-  Textarea,
-  Button,
-} from '@chakra-ui/react';
+import { chakra, Code, extendTheme, Text, theme as chakraTheme, UnorderedList, useColorModeValue } from '@chakra-ui/react';
 import { mode } from '@chakra-ui/theme-tools';
 import { AppSeoProps, CombinedThemeProvider, DocsPage, ExtendComponents, handlePushRoute } from '@guild-docs/client';
 import { Footer, Header, Subheader } from '@theguild/components';
-
 import type { AppProps } from 'next/app';
-import * as React from 'react';
-
-const envelopBase = () => /* TypeScript */ `
-import { envelop, useSchema } from "@envelop/core";
-import { schema } from "./schema";
-
-const getEnveloped = envelop({
- plugins: [useSchema(schema)],
- enableInternalTracing: false,
-});
-
-export { getEnveloped }
-`;
-
-const EnvelopPlayground = () => {
-  const [envelopCode, setEnvelopCode] = React.useState(envelopBase);
-  const [document, setDocument] = React.useState('query { __typename }');
-  const [result, setResult] = React.useState('');
-
-  const workerRef = React.useRef<Worker | null>(null);
-  React.useEffect(() => {
-    // @ts-ignore
-    workerRef.current = new Worker(new URL('../lib/worker.ts', import.meta.url));
-    workerRef.current.onmessage = evt => {
-      console.log(evt.data);
-      if (evt.data.type === 'execute-result') {
-        setResult(JSON.stringify(JSON.parse(evt.data.result), null, 2));
-      }
-    };
-    return () => {
-      workerRef.current?.terminate();
-    };
-  }, []);
-
-  React.useEffect(() => {}, [envelopCode]);
-
-  return (
-    <>
-      <Text mb="8px">Envelop code</Text>
-      <Textarea value={envelopCode} onChange={ev => setEnvelopCode(ev.target.value)} />
-      <Button onClick={() => workerRef.current?.postMessage({ id: '1', type: 'code', code: envelopCode })}>Bundle</Button>
-      <Text mb="8px">Document</Text>
-      <Textarea value={document} onChange={ev => setDocument(ev.target.value)} />
-      <Button
-        onClick={() => {
-          workerRef.current?.postMessage({ id: '1', type: 'execute', payload: { query: document } });
-        }}
-      >
-        Execute
-      </Button>
-      <Text>Result</Text>
-      <pre>
-        <code>{result}</code>
-      </pre>
-    </>
-  );
-};
+import { EnvelopPlayground } from '../components/envelop-playground';
 
 ExtendComponents({
   a: chakra('a', {
@@ -103,7 +36,7 @@ ExtendComponents({
   },
   Text,
   ul: UnorderedList,
-  EnvelopPlayground: () => <EnvelopPlayground />,
+  EnvelopPlayground,
 });
 
 const styles: typeof chakraTheme['styles'] = {
